Change `discreteSpread` to `spread`
